seperated stuff, added tests
