+++ torchao/prototype/intx/bitpacking.py
@@ -185,7 +185,6 @@ def pack(data: torch.Tensor,
     a less branching but more compute version so better for gpu
     '''
     container_size = 8
-    # torch._assert(data.shape[dim] % 8 == 0, f"pack dimension size ({data.shape[dim]}) is not divisble by scale")
     shards = [(data & maskbits[elem_size][i]) >> shifts[elem_size][i] for i in range(len(maskbits[elem_size]))]
     return tuple([_pack(shards[i], numbits[elem_size][i], container_size//numbits[elem_size][i], dim) for i in range(len(maskbits[elem_size]))])
